+++ un_yaml/__init__.py
@@ -1,3 +1,4 @@
 from .un_cli import UnCli  # NOQA F401
+from .un_conf import UnConf  # NOQA F401
 from .un_uri import UnUri  # NOQA F401
 from .un_yaml import UnYaml  # NOQA F401